+++ bin/client.py
@@ -18,11 +18,12 @@ handler.setFormatter(logging.Formatter('%(asctime)-15s [%(levelname)-8s] %(messa
 log.addHandler(handler)
 
 # setup ancp session
-client = Client(address="1.2.3.4")
+client = Client(address="172.30.138.10")
 if client.connect():
     # create ancp subscribers
-    S1 = Subscriber(aci="0.0.0.0 eth 1", up=1024, down=16000)
-    S2 = Subscriber(aci="0.0.0.0 eth 2", up=2048, down=32000)
+    S1 = Subscriber(aci="0.0.0.0 eth 1", up=1024, down=16000, aaci_bin=128, aaci_ascii="128")
+    S2 = Subscriber(aci="0.0.0.0 eth 2", up=2048, down=32000, aaci_bin=(128, 7), aaci_ascii="128")
+
     # send port-up for ancp subscribers
     client.port_up([S1, S2])
     # keep session active